Eduflex/requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-braces==1.8.1
django-crispy-forms==1.7.2
django-embed-video==1.0.0
django-pagedown==1.0.6
idna==2.7
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
